requirements.txt
flask==3.0.0
flask-cors==4.0.0
chromadb==0.4.22
requests==2.31.0
python-dotenv==1.0.0
numpy<2.0.0
